requirements.txt
Keras==2.2.4
tensorflow-gpu==1.14.0
h5py==2.10.0
pandas>=0.19.2
pillow>=2.1.0
requests
matplotlib>=2.0.0
opencv-python
seaborn
efficientnet
